+++ .gitignore
@@ -139,3 +139,4 @@ dmypy.json
 **/.DS_Store
 .aider*
 QualityScaler_temp
+src\qualityscaler\AI\
