+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "qualityscaler"
+name = "quality_scaler"
 description = "upscales images using ai"
 requires-python = ">=3.7"
 keywords = ["template-python-cmd"]
@@ -24,7 +24,7 @@ dependencies = [
     "Pillow",
     "download",
 ]
-version = "2.8.0"
+version = "2.12.0"
 # readme is labeled as dynamic
 [tool.setuptools.dynamic]
 dependencies = { file = ["requirements.txt"] }
@@ -55,4 +55,5 @@ ignore_missing_imports = true
 disable_error_code = ["import-untyped"]
 
 [project.scripts]
+quality-scaler = "qualityscaler.cli:main"
 qualityscaler = "qualityscaler.cli:main"
