+++ src/qualityscaler/assets/clear_icon.png
