+++ src/qualityscaler/assets/example.txt
