+++ src/qualityscaler/assets/github_logo.png
