+++ src/qualityscaler/assets/info_icon.png
