+++ src/qualityscaler/assets/logo.ico
