+++ src/qualityscaler/assets/logo.png
