+++ src/qualityscaler/assets/out.txt
