+++ src/qualityscaler/assets/stop_icon.png
