+++ src/qualityscaler/assets/telegram_logo.png
