+++ src/qualityscaler/assets/upscale_icon.png
