+++ src/qualityscaler/cli.py
@@ -11,7 +11,7 @@ def main() -> int:
     """Main entry point for the template_python_cmd package."""
     python_exe = sys.executable
     print(f"Python executable: {python_exe}")
-    rtn = os.system(f"{python_exe} -m quality_scaler.QualityScaler")
+    rtn = os.system(f"{python_exe} -m qualityscaler.QualityScaler")
     return rtn
